feat: useLoadedModelStore separated into two files
