+++ src/stores/LiveModel/LoadedModelStore/ModelState.ts
@@ -1,0 +1,8 @@
+import type { ModelType } from '../../../types';
+
+export type ModelState = {
+  loadedModelId: number;
+  loadedModelType: ModelType;
+  setLoadedModel: (id: number, type: ModelType) => void;
+  clear: () => void;
+};
